Found and fixed a number of bugs
Making convert.py modular introduced a number of bugs due to variables
not being used or going into the wrong places. Fixed a number of these.

In a break of compatibility with old code: 'modified_lines' no longer
contains tuples. 'original_lines' and 'modified_lines' are returned as
separate lists to improve readability and simplicity.

Have replaced the old 'os.system' with the new 'subprocess' module. This
captures stdout and raises exceptions if anything was passed to
stderr (ie if the ledger program returns an error).

Used mypy, flake8 and black for linting to improve code compliance with
the PEPs. I have also put in a number of type annotations that helped
find bugs.

Removed .gitignore from tracking so that everybody can have their own
gitignore file in their local repo. Recommend everybody adds
'.gitignore' to their .gitignore file so that it doesn't bother them.

I still have to do a full test of this modular branch with my real data
to confirm it is working correctly. That may show more bugs to fix.